+++ packages/gatsby-transformer-remark/src/on-node-create.js
@@ -4,14 +4,20 @@ const _ = require(`lodash`)
 
 module.exports = async function onCreateNode(
   { node, loadNodeContent, actions, createNodeId, reporter },
-  pluginOptions
+  {
+    plugins = null,
+    filter = () => true,
+    type = `MarkdownRemark`,
+    ...grayMatterOptions
+  } = {}
 ) {
   const { createNode, createParentChildLink } = actions
 
   // We only care about markdown content.
   if (
-    node.internal.mediaType !== `text/markdown` &&
-    node.internal.mediaType !== `text/x-markdown`
+    (node.internal.mediaType !== `text/markdown` &&
+      node.internal.mediaType !== `text/x-markdown`) ||
+    !filter(node)
   ) {
     return {}
   }
@@ -19,7 +25,7 @@ module.exports = async function onCreateNode(
   const content = await loadNodeContent(node)
 
   try {
-    let data = grayMatter(content, pluginOptions)
+    let data = grayMatter(content, grayMatterOptions)
 
     if (data.data) {
       data.data = _.mapValues(data.data, value => {
@@ -31,12 +37,12 @@ module.exports = async function onCreateNode(
     }
 
     let markdownNode = {
-      id: createNodeId(`${node.id} >>> MarkdownRemark`),
+      id: createNodeId(`${node.id} >>> ${type}`),
       children: [],
       parent: node.id,
       internal: {
         content: data.content,
-        type: `MarkdownRemark`,
+        type,
       },
     }