+++ www/src/components/navigation.js
@@ -5,13 +5,13 @@ import TwitterIcon from "react-icons/lib/fa/twitter"
 import SearchForm from "../components/search-form"
 import DiscordIcon from "../components/discord"
 import logo from "../logo.svg"
-import { rhythm, scale, options } from "../utils/typography"
-import presets, { colors } from "../utils/presets"
+import { rhythm, options } from "../utils/typography"
+import presets, { colors, space, transition } from "../utils/presets"
 
 // what we need to nudge down the navItems to sit
 // on the baseline of the logo's wordmark
 const navItemTopOffset = `0.6rem`
-const navItemHorizontalSpacing = rhythm(1 / 3)
+const navItemHorizontalSpacing = rhythm(space[2])
 
 const iconColor = colors.lilac
 
@@ -142,8 +142,7 @@ const Navigation = ({ pathname }) => {
           <div
             css={{
               display: `none`,
-              [presets.Lg]: { display: !isHomepage && `flex` },
-              [presets.Xl]: { display: `flex` },
+              [presets.Lg]: { display: `flex` },
             }}
           >
             <SocialNavItem href="https://gatsby.dev/discord" title="Discord">
@@ -187,7 +186,6 @@ const styles = {
       display: `flex`,
       flexGrow: 1,
       margin: 0,
-      marginLeft: rhythm(1 / 4),
       listStyle: `none`,
       maskImage: `linear-gradient(to right, transparent, white ${rhythm(
         1 / 8
@@ -197,8 +195,8 @@ const styles = {
   },
   containerInner: {
     margin: `0 auto`,
-    paddingLeft: rhythm(3 / 4),
-    paddingRight: rhythm(3 / 4),
+    paddingLeft: rhythm(space[6]),
+    paddingRight: rhythm(space[6]),
     fontFamily: options.headerFontFamily.join(`,`),
     display: `flex`,
     alignItems: `center`,
@@ -206,20 +204,18 @@ const styles = {
     height: `100%`,
   },
   navItem: {
-    ...scale(-1 / 3),
+    fontSize: presets.scale[1],
     borderBottom: `0.125rem solid transparent`,
     color: `inherit`,
     display: `block`,
-    letterSpacing: `0.03em`,
+    letterSpacing: presets.letterSpacings.tracked,
     WebkitFontSmoothing: `antialiased`,
     lineHeight: `calc(${presets.headerHeight} - ${navItemTopOffset})`,
     position: `relative`,
     textDecoration: `none`,
     textTransform: `uppercase`,
     top: 0,
-    transition: `color ${presets.animation.speedDefault} ${
-      presets.animation.curveDefault
-    }`,
+    transition: `color ${transition.speed.default} ${transition.curve.default}`,
     zIndex: 1,
     "&:hover": {
       color: colors.gatsby,
@@ -240,18 +236,15 @@ const styles = {
     marginLeft: `auto`,
   },
   logo: {
-    height: 28,
+    height: rhythm(space[6]),
     margin: 0,
-    [presets.Md]: {
-      height: `1.55rem`,
-    },
   },
   logoLink: {
     alignItems: `center`,
     color: `inherit`,
     display: `flex`,
     flexShrink: 0,
-    marginRight: rhythm(1 / 2),
+    marginRight: rhythm(space[3]),
     textDecoration: `none`,
   },
 }