+++ www/src/components/plugin-searchbar-body.js
@@ -9,7 +9,6 @@ import {
   Toggle,
 } from "react-instantsearch/dom"
 import { navigate as reachNavigate } from "@reach/router"
-import { colors } from "../utils/presets"
 import { Link } from "gatsby"
 import DownloadArrow from "react-icons/lib/md/file-download"
 import AlgoliaLogo from "../assets/algolia.svg"
@@ -17,9 +16,9 @@ import GatsbyIcon from "../monogram.svg"
 import debounce from "lodash/debounce"
 import unescape from "lodash/unescape"
 
-import presets from "../utils/presets"
-import typography, { rhythm, scale } from "../utils/typography"
-import { scrollbarStyles } from "../utils/styles"
+import presets, { space, colors, transition, radii } from "../utils/presets"
+import { rhythm, options } from "../utils/typography"
+import { scrollbarStyles, skipLink } from "../utils/styles"
 import { Global, css } from "@emotion/core"
 import styled from "@emotion/styled"
 import removeMD from "remove-markdown"
@@ -32,7 +31,7 @@ const updateAfter = 700
 // A couple constants for CSS
 const searchInputHeight = rhythm(7 / 4)
 const searchMetaHeight = rhythm(8 / 4)
-const searchInputWrapperMargin = rhythm(3 / 4)
+const searchInputWrapperMargin = rhythm(space[6])
 
 /* stylelint-disable */
 const searchBoxStyles = css`
@@ -58,20 +57,21 @@ const searchBoxStyles = css`
 
   .ais-SearchBox__input {
     -webkit-appearance: none;
-    background: #fff;
+    background: ${colors.white};
     border: 1px solid ${colors.ui.bright};
-    border-radius: ${presets.radiusLg}px;
+    border-radius: ${radii[2]}px;
     color: ${colors.gatsby};
     display: inline-block;
-    font-size: 18px;
-    font-family: ${typography.options.headerFontFamily.join(`,`)};
+    font-size: ${presets.scale[3]};
+    font-family: ${options.headerFontFamily.join(`,`)};
     height: ${searchInputHeight};
     padding: 0;
     padding-right: ${searchInputHeight};
     padding-left: ${searchInputHeight};
     margin: 0 ${searchInputWrapperMargin};
-    -webkit-transition: box-shadow 0.4s ease, background 0.4s ease;
-    transition: box-shadow 0.4s ease, background 0.4s ease;
+    transition: box-shadow ${transition.speed.default}
+        ${transition.curve.default},
+      background ${transition.speed.default} ${transition.curve.default};
     vertical-align: middle;
     white-space: normal;
     width: calc(100% - ${rhythm(6 / 4)});
@@ -87,7 +87,6 @@ const searchBoxStyles = css`
   .ais-SearchBox__input:focus {
     border-color: ${colors.lilac};
     box-shadow: 0 0 0 3px ${colors.ui.bright};
-    transition: border-color 0.15s ease-in-out, box-shadow 0.15s ease-in-out;
   }
 
   .ais-SearchBox__input::-webkit-input-placeholder,
@@ -119,7 +118,7 @@ const searchBoxStyles = css`
     top: ${searchInputWrapperMargin};
     right: inherit;
     left: ${searchInputWrapperMargin};
-    border-radius: ${presets.radiusLg}px 0 0 ${presets.radiusLg}px;
+    border-radius: ${radii[2]}px 0 0 ${radii[2]}px;
   }
   .ais-SearchBox__submit:focus {
     outline: 0;
@@ -128,8 +127,8 @@ const searchBoxStyles = css`
     fill: ${colors.gatsby};
   }
   .ais-SearchBox__submit svg {
-    width: 1rem;
-    height: 1rem;
+    width: ${rhythm(space[4])};
+    height: ${rhythm(space[4])};
     vertical-align: middle;
     fill: ${colors.ui.bright};
   }
@@ -158,21 +157,20 @@ const searchBoxStyles = css`
   .ais-InfiniteHits__loadMore {
     background-color: transparent;
     border: 1px solid ${colors.gatsby};
-    border-radius: ${presets.radius}px;
+    border-radius: ${radii[1]}px;
     color: ${colors.gatsby};
     cursor: pointer;
-    width: calc(100% - ${rhythm(6 / 4)});
-    margin: ${rhythm(3 / 4)};
-    height: ${rhythm(2)};
+    width: calc(100% - ${rhythm(space[6] * 2)});
+    margin: ${rhythm(space[6])};
+    height: ${rhythm(space[9])};
     outline: none;
-    transition: all ${presets.animation.speedDefault}
-      ${presets.animation.curveDefault};
-    font-family: ${typography.options.headerFontFamily.join(`,`)};
+    transition: all ${transition.speed.default} ${transition.curve.default};
+    font-family: ${options.headerFontFamily.join(`,`)};
   }
   .ais-InfiniteHits__loadMore:hover,
   .ais-InfiniteHits__loadMore:focus {
     background-color: ${colors.gatsby};
-    color: #fff;
+    color: ${colors.white};
   }
 
   .ais-InfiniteHits__loadMore[disabled] {
@@ -181,29 +179,7 @@ const searchBoxStyles = css`
 `
 /* stylelint-enable */
 
-const StyledSkipNavLink = styled(SkipNavLink)`
-  border: 0;
-  clip: rect(0 0 0 0);
-  height: 1px;
-  width: 1px;
-  margin: -1px;
-  padding: 0;
-  overflow: hidden;
-  position: absolute;
-  z-index: 100;
-  font-size: 0.85rem;
-
-  :focus {
-    padding: 0.9rem;
-    top: 10px;
-    left: 10px;
-    background: white;
-    text-decoration: none;
-    width: auto;
-    height: auto;
-    clip: auto;
-  }
-`
+const StyledSkipNavLink = styled(SkipNavLink)({ ...skipLink })
 
 // Search shows a list of "hits", and is a child of the PluginSearchBar component
 class Search extends Component {
@@ -232,6 +208,14 @@ class Search extends Component {
             <Configure analyticsTags={[`gatsby-plugins`]} />
             <RefinementList
               attributeName="keywords"
+              transformItems={items =>
+                items.map(({ count, ...item }) => {
+                  return {
+                    ...item,
+                    count: count || 0,
+                  }
+                })
+              }
               defaultRefinement={[`gatsby-component`, `gatsby-plugin`]}
             />
             <Toggle
@@ -247,14 +231,10 @@ class Search extends Component {
               alignItems: `center`,
               color: colors.gray.calm,
               display: `flex`,
-              fontFamily: typography.options.headerFontFamily.join(`,`),
-              fontSize: scale(1),
               height: searchMetaHeight,
-              paddingLeft: rhythm(3 / 4),
-              paddingRight: rhythm(3 / 4),
-              [presets.Md]: {
-                fontSize: scale(-1 / 4).fontSize,
-              },
+              paddingLeft: rhythm(space[6]),
+              paddingRight: rhythm(space[6]),
+              fontSize: presets.scale[1],
             }}
           >
             <Stats
@@ -297,7 +277,7 @@ class Search extends Component {
             fontSize: 0,
             lineHeight: 0,
             height: 20,
-            marginTop: rhythm(3 / 4),
+            marginTop: rhythm(space[6]),
             display: `none`,
           }}
         >
@@ -307,7 +287,6 @@ class Search extends Component {
               "&&": {
                 background: `url(${AlgoliaLogo})`,
                 border: `none`,
-                boxShadow: `none`,
                 fontWeight: `normal`,
                 backgroundRepeat: `no-repeat`,
                 backgroundPosition: `50%`,
@@ -347,20 +326,19 @@ const Result = ({ hit, pathname, query }) => {
       to={`/packages/${hit.name}/?=${query}`}
       css={{
         "&&": {
-          boxShadow: `none`,
-          background: selected ? `#fff` : false,
+          background: selected ? colors.white : false,
           borderBottom: 0,
           color: colors.gray.dark,
           display: `block`,
           fontWeight: `400`,
-          padding: rhythm(3 / 4),
+          padding: `${rhythm(space[4])} ${rhythm(space[6])}`,
           position: `relative`,
-          transition: `all ${presets.animation.speedDefault} ${
-            presets.animation.curveDefault
+          transition: `all ${transition.speed.default} ${
+            transition.curve.default
           }`,
           zIndex: selected ? 1 : false,
           "&:hover": {
-            background: selected ? `#fff` : colors.ui.border,
+            background: selected ? colors.white : colors.ui.border,
           },
           "&:before": {
             background: colors.ui.border,
@@ -389,21 +367,19 @@ const Result = ({ hit, pathname, query }) => {
           alignItems: `baseline`,
           display: `flex`,
           justifyContent: `space-between`,
-          marginBottom: rhythm(typography.options.blockMarginBottom / 2),
+          marginBottom: rhythm(space[3]),
         }}
       >
         <h2
           css={{
             color: selected ? colors.gatsby : false,
             fontSize: `inherit`,
-            fontFamily: typography.options.headerFontFamily.join(`,`),
+            fontFamily: options.headerFontFamily.join(`,`),
             fontWeight: `bold`,
-            lineHeight: 1.2,
             display: `flex`,
             alignItems: `center`,
             marginBottom: 0,
             marginTop: 0,
-            letterSpacing: 0,
           }}
         >
           {hit.name}
@@ -419,8 +395,7 @@ const Result = ({ hit, pathname, query }) => {
             alignItems: `center`,
             color: selected ? colors.lilac : colors.gray.bright,
             display: `flex`,
-            fontFamily: typography.options.headerFontFamily.join(`,`),
-            fontSize: rhythm(4 / 7),
+            fontSize: presets.scale[0],
           }}
         >
           {hit.repository &&
@@ -441,7 +416,7 @@ const Result = ({ hit, pathname, query }) => {
             )}
           <span
             css={{
-              width: `4.5em`,
+              width: `5em`,
               textAlign: `right`,
             }}
           >
@@ -450,7 +425,7 @@ const Result = ({ hit, pathname, query }) => {
             <span
               css={{
                 color: selected ? colors.lilac : colors.gray.bright,
-                marginLeft: rhythm(1 / 6),
+                marginLeft: rhythm(space[1]),
               }}
             >
               <DownloadArrow />
@@ -461,9 +436,7 @@ const Result = ({ hit, pathname, query }) => {
       <div
         css={{
           color: selected ? `inherit` : colors.gray.calm,
-          fontFamily: typography.options.systemFontFamily.join(`,`),
-          fontSize: scale(-1 / 2).fontSize,
-          lineHeight: 1.5,
+          fontSize: presets.scale[1],
         }}
       >
         {removeMD(unescape(hit.description))}