+++ packages/gatsby-plugin-remove-trailing-slashes/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.10](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-remove-trailing-slashes/compare/gatsby-plugin-remove-trailing-slashes@2.0.9...gatsby-plugin-remove-trailing-slashes@2.0.10) (2019-03-11)
+
+**Note:** Version bump only for package gatsby-plugin-remove-trailing-slashes
+
 ## [2.0.9](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-remove-trailing-slashes/compare/gatsby-plugin-remove-trailing-slashes@2.0.8...gatsby-plugin-remove-trailing-slashes@2.0.9) (2019-02-28)
 
 **Note:** Version bump only for package gatsby-plugin-remove-trailing-slashes