+++ packages/gatsby-transformer-documentationjs/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-documentationjs/compare/gatsby-transformer-documentationjs@2.0.4...gatsby-transformer-documentationjs@2.0.5) (2019-03-11)
+
+**Note:** Version bump only for package gatsby-transformer-documentationjs
+
 ## [2.0.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-transformer-documentationjs/compare/gatsby-transformer-documentationjs@2.0.3...gatsby-transformer-documentationjs@2.0.4) (2019-02-06)
 
 **Note:** Version bump only for package gatsby-transformer-documentationjs